requirements.txt
aiohttp[speedups]
numpy
opencv-python
onnxruntime-gpu
cuda-python
tensorrt
PyYAML